Functional mapping with optional parallelism and progress bars

Applies a function `.f` to each element of `.x`, with optional parallel processing and progress bar support.

Usage

fmap(.x, .f, ncores = NULL, pb = FALSE, ...)

Arguments

.x

A list or atomic vector of elements to iterate over.

.f

A function to apply to each element of `.x`. Can be a function or a string naming a function.

ncores

Integer. Number of CPU cores to use for parallel processing. Default is `NULL` (sequential).

pb

Logical. Whether to show a progress bar. Default is `FALSE`.

...

Additional arguments passed to `.f`.

Value

A list of results, one for each element of `.x`.

Examples

slow_fn <- function(x) { Sys.sleep(0.01); x^2 }
x <- 1:100

# Basic usage
fmap(x, slow_fn)
#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 4
#> 
#> [[3]]
#> [1] 9
#> 
#> [[4]]
#> [1] 16
#> 
#> [[5]]
#> [1] 25
#> 
#> [[6]]
#> [1] 36
#> 
#> [[7]]
#> [1] 49
#> 
#> [[8]]
#> [1] 64
#> 
#> [[9]]
#> [1] 81
#> 
#> [[10]]
#> [1] 100
#> 
#> [[11]]
#> [1] 121
#> 
#> [[12]]
#> [1] 144
#> 
#> [[13]]
#> [1] 169
#> 
#> [[14]]
#> [1] 196
#> 
#> [[15]]
#> [1] 225
#> 
#> [[16]]
#> [1] 256
#> 
#> [[17]]
#> [1] 289
#> 
#> [[18]]
#> [1] 324
#> 
#> [[19]]
#> [1] 361
#> 
#> [[20]]
#> [1] 400
#> 
#> [[21]]
#> [1] 441
#> 
#> [[22]]
#> [1] 484
#> 
#> [[23]]
#> [1] 529
#> 
#> [[24]]
#> [1] 576
#> 
#> [[25]]
#> [1] 625
#> 
#> [[26]]
#> [1] 676
#> 
#> [[27]]
#> [1] 729
#> 
#> [[28]]
#> [1] 784
#> 
#> [[29]]
#> [1] 841
#> 
#> [[30]]
#> [1] 900
#> 
#> [[31]]
#> [1] 961
#> 
#> [[32]]
#> [1] 1024
#> 
#> [[33]]
#> [1] 1089
#> 
#> [[34]]
#> [1] 1156
#> 
#> [[35]]
#> [1] 1225
#> 
#> [[36]]
#> [1] 1296
#> 
#> [[37]]
#> [1] 1369
#> 
#> [[38]]
#> [1] 1444
#> 
#> [[39]]
#> [1] 1521
#> 
#> [[40]]
#> [1] 1600
#> 
#> [[41]]
#> [1] 1681
#> 
#> [[42]]
#> [1] 1764
#> 
#> [[43]]
#> [1] 1849
#> 
#> [[44]]
#> [1] 1936
#> 
#> [[45]]
#> [1] 2025
#> 
#> [[46]]
#> [1] 2116
#> 
#> [[47]]
#> [1] 2209
#> 
#> [[48]]
#> [1] 2304
#> 
#> [[49]]
#> [1] 2401
#> 
#> [[50]]
#> [1] 2500
#> 
#> [[51]]
#> [1] 2601
#> 
#> [[52]]
#> [1] 2704
#> 
#> [[53]]
#> [1] 2809
#> 
#> [[54]]
#> [1] 2916
#> 
#> [[55]]
#> [1] 3025
#> 
#> [[56]]
#> [1] 3136
#> 
#> [[57]]
#> [1] 3249
#> 
#> [[58]]
#> [1] 3364
#> 
#> [[59]]
#> [1] 3481
#> 
#> [[60]]
#> [1] 3600
#> 
#> [[61]]
#> [1] 3721
#> 
#> [[62]]
#> [1] 3844
#> 
#> [[63]]
#> [1] 3969
#> 
#> [[64]]
#> [1] 4096
#> 
#> [[65]]
#> [1] 4225
#> 
#> [[66]]
#> [1] 4356
#> 
#> [[67]]
#> [1] 4489
#> 
#> [[68]]
#> [1] 4624
#> 
#> [[69]]
#> [1] 4761
#> 
#> [[70]]
#> [1] 4900
#> 
#> [[71]]
#> [1] 5041
#> 
#> [[72]]
#> [1] 5184
#> 
#> [[73]]
#> [1] 5329
#> 
#> [[74]]
#> [1] 5476
#> 
#> [[75]]
#> [1] 5625
#> 
#> [[76]]
#> [1] 5776
#> 
#> [[77]]
#> [1] 5929
#> 
#> [[78]]
#> [1] 6084
#> 
#> [[79]]
#> [1] 6241
#> 
#> [[80]]
#> [1] 6400
#> 
#> [[81]]
#> [1] 6561
#> 
#> [[82]]
#> [1] 6724
#> 
#> [[83]]
#> [1] 6889
#> 
#> [[84]]
#> [1] 7056
#> 
#> [[85]]
#> [1] 7225
#> 
#> [[86]]
#> [1] 7396
#> 
#> [[87]]
#> [1] 7569
#> 
#> [[88]]
#> [1] 7744
#> 
#> [[89]]
#> [1] 7921
#> 
#> [[90]]
#> [1] 8100
#> 
#> [[91]]
#> [1] 8281
#> 
#> [[92]]
#> [1] 8464
#> 
#> [[93]]
#> [1] 8649
#> 
#> [[94]]
#> [1] 8836
#> 
#> [[95]]
#> [1] 9025
#> 
#> [[96]]
#> [1] 9216
#> 
#> [[97]]
#> [1] 9409
#> 
#> [[98]]
#> [1] 9604
#> 
#> [[99]]
#> [1] 9801
#> 
#> [[100]]
#> [1] 10000
#> 

# With progress bar
fmap(x, slow_fn, pb = TRUE)
#>  |                                                  |   0% elapsed=00h 00m 00s, remaining~... |                                                  |   1% elapsed=00h 00m 00s, remaining~00h 00m 01s |=                                                 |   2% elapsed=00h 00m 00s, remaining~00h 00m 01s |=                                                 |   3% elapsed=00h 00m 00s, remaining~00h 00m 01s |==                                                |   4% elapsed=00h 00m 00s, remaining~00h 00m 01s |==                                                |   5% elapsed=00h 00m 00s, remaining~00h 00m 01s |===                                               |   6% elapsed=00h 00m 00s, remaining~00h 00m 01s |===                                               |   7% elapsed=00h 00m 00s, remaining~00h 00m 01s |====                                              |   8% elapsed=00h 00m 00s, remaining~00h 00m 01s |====                                              |   9% elapsed=00h 00m 00s, remaining~00h 00m 01s |=====                                             |  10% elapsed=00h 00m 00s, remaining~00h 00m 01s |=====                                             |  11% elapsed=00h 00m 00s, remaining~00h 00m 01s |======                                            |  12% elapsed=00h 00m 00s, remaining~00h 00m 01s |======                                            |  13% elapsed=00h 00m 00s, remaining~00h 00m 01s |=======                                           |  14% elapsed=00h 00m 00s, remaining~00h 00m 01s |=======                                           |  15% elapsed=00h 00m 00s, remaining~00h 00m 01s |========                                          |  16% elapsed=00h 00m 00s, remaining~00h 00m 01s |========                                          |  17% elapsed=00h 00m 00s, remaining~00h 00m 01s |=========                                         |  18% elapsed=00h 00m 00s, remaining~00h 00m 01s |=========                                         |  19% elapsed=00h 00m 00s, remaining~00h 00m 01s |==========                                        |  20% elapsed=00h 00m 00s, remaining~00h 00m 01s |==========                                        |  21% elapsed=00h 00m 00s, remaining~00h 00m 01s |===========                                       |  22% elapsed=00h 00m 00s, remaining~00h 00m 01s |===========                                       |  23% elapsed=00h 00m 00s, remaining~00h 00m 01s |============                                      |  24% elapsed=00h 00m 00s, remaining~00h 00m 01s |============                                      |  25% elapsed=00h 00m 00s, remaining~00h 00m 01s |=============                                     |  26% elapsed=00h 00m 00s, remaining~00h 00m 01s |=============                                     |  27% elapsed=00h 00m 00s, remaining~00h 00m 01s |==============                                    |  28% elapsed=00h 00m 00s, remaining~00h 00m 01s |==============                                    |  29% elapsed=00h 00m 00s, remaining~00h 00m 01s |===============                                   |  30% elapsed=00h 00m 00s, remaining~00h 00m 01s |===============                                   |  31% elapsed=00h 00m 00s, remaining~00h 00m 01s |================                                  |  32% elapsed=00h 00m 00s, remaining~00h 00m 01s |================                                  |  33% elapsed=00h 00m 00s, remaining~00h 00m 01s |=================                                 |  34% elapsed=00h 00m 00s, remaining~00h 00m 01s |=================                                 |  35% elapsed=00h 00m 00s, remaining~00h 00m 01s |==================                                |  36% elapsed=00h 00m 00s, remaining~00h 00m 01s |==================                                |  37% elapsed=00h 00m 00s, remaining~00h 00m 01s |===================                               |  38% elapsed=00h 00m 00s, remaining~00h 00m 01s |===================                               |  39% elapsed=00h 00m 00s, remaining~00h 00m 01s |====================                              |  40% elapsed=00h 00m 00s, remaining~00h 00m 01s |====================                              |  41% elapsed=00h 00m 00s, remaining~00h 00m 01s |=====================                             |  42% elapsed=00h 00m 00s, remaining~00h 00m 01s |=====================                             |  43% elapsed=00h 00m 00s, remaining~00h 00m 01s |======================                            |  44% elapsed=00h 00m 00s, remaining~00h 00m 01s |======================                            |  45% elapsed=00h 00m 00s, remaining~00h 00m 01s |=======================                           |  46% elapsed=00h 00m 00s, remaining~00h 00m 01s |=======================                           |  47% elapsed=00h 00m 00s, remaining~00h 00m 01s |========================                          |  48% elapsed=00h 00m 01s, remaining~00h 00m 01s |========================                          |  49% elapsed=00h 00m 01s, remaining~00h 00m 01s |=========================                         |  50% elapsed=00h 00m 01s, remaining~00h 00m 01s |=========================                         |  51% elapsed=00h 00m 01s, remaining~00h 00m 01s |==========================                        |  52% elapsed=00h 00m 01s, remaining~00h 00m 01s |==========================                        |  53% elapsed=00h 00m 01s, remaining~00h 00m 00s |===========================                       |  54% elapsed=00h 00m 01s, remaining~00h 00m 00s |===========================                       |  55% elapsed=00h 00m 01s, remaining~00h 00m 00s |============================                      |  56% elapsed=00h 00m 01s, remaining~00h 00m 00s |============================                      |  57% elapsed=00h 00m 01s, remaining~00h 00m 00s |============================                      |  58% elapsed=00h 00m 01s, remaining~00h 00m 00s |=============================                     |  59% elapsed=00h 00m 01s, remaining~00h 00m 00s |==============================                    |  60% elapsed=00h 00m 01s, remaining~00h 00m 00s |==============================                    |  61% elapsed=00h 00m 01s, remaining~00h 00m 00s |===============================                   |  62% elapsed=00h 00m 01s, remaining~00h 00m 00s |===============================                   |  63% elapsed=00h 00m 01s, remaining~00h 00m 00s |================================                  |  64% elapsed=00h 00m 01s, remaining~00h 00m 00s |================================                  |  65% elapsed=00h 00m 01s, remaining~00h 00m 00s |=================================                 |  66% elapsed=00h 00m 01s, remaining~00h 00m 00s |=================================                 |  67% elapsed=00h 00m 01s, remaining~00h 00m 00s |==================================                |  68% elapsed=00h 00m 01s, remaining~00h 00m 00s |==================================                |  69% elapsed=00h 00m 01s, remaining~00h 00m 00s |===================================               |  70% elapsed=00h 00m 01s, remaining~00h 00m 00s |===================================               |  71% elapsed=00h 00m 01s, remaining~00h 00m 00s |====================================              |  72% elapsed=00h 00m 01s, remaining~00h 00m 00s |====================================              |  73% elapsed=00h 00m 01s, remaining~00h 00m 00s |=====================================             |  74% elapsed=00h 00m 01s, remaining~00h 00m 00s |=====================================             |  75% elapsed=00h 00m 01s, remaining~00h 00m 00s |======================================            |  76% elapsed=00h 00m 01s, remaining~00h 00m 00s |======================================            |  77% elapsed=00h 00m 01s, remaining~00h 00m 00s |=======================================           |  78% elapsed=00h 00m 01s, remaining~00h 00m 00s |=======================================           |  79% elapsed=00h 00m 01s, remaining~00h 00m 00s |========================================          |  80% elapsed=00h 00m 01s, remaining~00h 00m 00s |========================================          |  81% elapsed=00h 00m 01s, remaining~00h 00m 00s |=========================================         |  82% elapsed=00h 00m 01s, remaining~00h 00m 00s |=========================================         |  83% elapsed=00h 00m 01s, remaining~00h 00m 00s |==========================================        |  84% elapsed=00h 00m 01s, remaining~00h 00m 00s |==========================================        |  85% elapsed=00h 00m 01s, remaining~00h 00m 00s |===========================================       |  86% elapsed=00h 00m 01s, remaining~00h 00m 00s |===========================================       |  87% elapsed=00h 00m 01s, remaining~00h 00m 00s |============================================      |  88% elapsed=00h 00m 01s, remaining~00h 00m 00s |============================================      |  89% elapsed=00h 00m 01s, remaining~00h 00m 00s |=============================================     |  90% elapsed=00h 00m 01s, remaining~00h 00m 00s |=============================================     |  91% elapsed=00h 00m 01s, remaining~00h 00m 00s |==============================================    |  92% elapsed=00h 00m 01s, remaining~00h 00m 00s |==============================================    |  93% elapsed=00h 00m 01s, remaining~00h 00m 00s |===============================================   |  94% elapsed=00h 00m 01s, remaining~00h 00m 00s |===============================================   |  95% elapsed=00h 00m 01s, remaining~00h 00m 00s |================================================  |  96% elapsed=00h 00m 01s, remaining~00h 00m 00s |================================================  |  97% elapsed=00h 00m 01s, remaining~00h 00m 00s |================================================= |  98% elapsed=00h 00m 01s, remaining~00h 00m 00s |================================================= |  99% elapsed=00h 00m 01s, remaining~00h 00m 00s |==================================================| 100% elapsed=00h 00m 01s, remaining~00h 00m 00s
#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 4
#> 
#> [[3]]
#> [1] 9
#> 
#> [[4]]
#> [1] 16
#> 
#> [[5]]
#> [1] 25
#> 
#> [[6]]
#> [1] 36
#> 
#> [[7]]
#> [1] 49
#> 
#> [[8]]
#> [1] 64
#> 
#> [[9]]
#> [1] 81
#> 
#> [[10]]
#> [1] 100
#> 
#> [[11]]
#> [1] 121
#> 
#> [[12]]
#> [1] 144
#> 
#> [[13]]
#> [1] 169
#> 
#> [[14]]
#> [1] 196
#> 
#> [[15]]
#> [1] 225
#> 
#> [[16]]
#> [1] 256
#> 
#> [[17]]
#> [1] 289
#> 
#> [[18]]
#> [1] 324
#> 
#> [[19]]
#> [1] 361
#> 
#> [[20]]
#> [1] 400
#> 
#> [[21]]
#> [1] 441
#> 
#> [[22]]
#> [1] 484
#> 
#> [[23]]
#> [1] 529
#> 
#> [[24]]
#> [1] 576
#> 
#> [[25]]
#> [1] 625
#> 
#> [[26]]
#> [1] 676
#> 
#> [[27]]
#> [1] 729
#> 
#> [[28]]
#> [1] 784
#> 
#> [[29]]
#> [1] 841
#> 
#> [[30]]
#> [1] 900
#> 
#> [[31]]
#> [1] 961
#> 
#> [[32]]
#> [1] 1024
#> 
#> [[33]]
#> [1] 1089
#> 
#> [[34]]
#> [1] 1156
#> 
#> [[35]]
#> [1] 1225
#> 
#> [[36]]
#> [1] 1296
#> 
#> [[37]]
#> [1] 1369
#> 
#> [[38]]
#> [1] 1444
#> 
#> [[39]]
#> [1] 1521
#> 
#> [[40]]
#> [1] 1600
#> 
#> [[41]]
#> [1] 1681
#> 
#> [[42]]
#> [1] 1764
#> 
#> [[43]]
#> [1] 1849
#> 
#> [[44]]
#> [1] 1936
#> 
#> [[45]]
#> [1] 2025
#> 
#> [[46]]
#> [1] 2116
#> 
#> [[47]]
#> [1] 2209
#> 
#> [[48]]
#> [1] 2304
#> 
#> [[49]]
#> [1] 2401
#> 
#> [[50]]
#> [1] 2500
#> 
#> [[51]]
#> [1] 2601
#> 
#> [[52]]
#> [1] 2704
#> 
#> [[53]]
#> [1] 2809
#> 
#> [[54]]
#> [1] 2916
#> 
#> [[55]]
#> [1] 3025
#> 
#> [[56]]
#> [1] 3136
#> 
#> [[57]]
#> [1] 3249
#> 
#> [[58]]
#> [1] 3364
#> 
#> [[59]]
#> [1] 3481
#> 
#> [[60]]
#> [1] 3600
#> 
#> [[61]]
#> [1] 3721
#> 
#> [[62]]
#> [1] 3844
#> 
#> [[63]]
#> [1] 3969
#> 
#> [[64]]
#> [1] 4096
#> 
#> [[65]]
#> [1] 4225
#> 
#> [[66]]
#> [1] 4356
#> 
#> [[67]]
#> [1] 4489
#> 
#> [[68]]
#> [1] 4624
#> 
#> [[69]]
#> [1] 4761
#> 
#> [[70]]
#> [1] 4900
#> 
#> [[71]]
#> [1] 5041
#> 
#> [[72]]
#> [1] 5184
#> 
#> [[73]]
#> [1] 5329
#> 
#> [[74]]
#> [1] 5476
#> 
#> [[75]]
#> [1] 5625
#> 
#> [[76]]
#> [1] 5776
#> 
#> [[77]]
#> [1] 5929
#> 
#> [[78]]
#> [1] 6084
#> 
#> [[79]]
#> [1] 6241
#> 
#> [[80]]
#> [1] 6400
#> 
#> [[81]]
#> [1] 6561
#> 
#> [[82]]
#> [1] 6724
#> 
#> [[83]]
#> [1] 6889
#> 
#> [[84]]
#> [1] 7056
#> 
#> [[85]]
#> [1] 7225
#> 
#> [[86]]
#> [1] 7396
#> 
#> [[87]]
#> [1] 7569
#> 
#> [[88]]
#> [1] 7744
#> 
#> [[89]]
#> [1] 7921
#> 
#> [[90]]
#> [1] 8100
#> 
#> [[91]]
#> [1] 8281
#> 
#> [[92]]
#> [1] 8464
#> 
#> [[93]]
#> [1] 8649
#> 
#> [[94]]
#> [1] 8836
#> 
#> [[95]]
#> [1] 9025
#> 
#> [[96]]
#> [1] 9216
#> 
#> [[97]]
#> [1] 9409
#> 
#> [[98]]
#> [1] 9604
#> 
#> [[99]]
#> [1] 9801
#> 
#> [[100]]
#> [1] 10000
#> 

# With parallel execution (non-Windows)
if (FALSE) { # \dontrun{
fmap(x, slow_fn, ncores = 4, pb = TRUE)
} # }

# Benchmark against furrr
if (FALSE) { # \dontrun{
library(furrr)
library(bench)
library(future)

plan(multisession, workers = 8)
bench::mark(
  funr_fmap    = fmap(x, slow_fn, ncores = 8, pb = FALSE),
  furrr_future = furrr::future_map(x, slow_fn),
  iterations = 10
)
} # }