Repeat an expression or function call multiple times

Repeats an expression or function evaluation `times` times. If `expr` is a function, it is invoked with optional input `.x` and additional arguments. If `expr` is a quoted expression, it is evaluated in the parent environment. Supports parallel processing and optional simplification of results.

Usage

frepeat(
  .x = NULL,
  times,
  expr,
  simplify = FALSE,
  ncores = NULL,
  pb = FALSE,
  ...
)

Arguments

.x

Optional input passed to `expr` if `expr` is a function. Default is `NULL`.

times

Integer. Number of repetitions.

expr

A function or an unevaluated expression to repeat. If a function, it will be called `times` times.

simplify

Logical. If `TRUE`, attempts to simplify the result using `simplify2array()`. Default is `FALSE`.

ncores

Integer. Number of cores to use for parallel execution. Default is `NULL` (sequential).

pb

Logical. Whether to display a progress bar. Default is `FALSE`.

...

Additional arguments passed to the function `expr` if it is callable.

Value

A list of outputs (or a simplified array if `simplify = TRUE`) from evaluating `expr` multiple times.

Note

If `expr` is passed as a function call (not a function or quoted expression), it will be evaluated immediately, not repeated. Use `function(...) \ ... \` or `quote(...)` instead..

Examples

# Repeat a pure function call
frepeat(times = 3, expr = function() rnorm(1))
#> [[1]]
#> [1] 0.3206011
#> 
#> [[2]]
#> [1] -1.369776
#> 
#> [[3]]
#> [1] 0.2912297
#> 

# Repeat a function with input `.x`
frepeat(.x = 10, times = 3, expr = function(x) rnorm(1, mean = x))
#> [[1]]
#> [1] 11.38555
#> 
#> [[2]]
#> [1] 10.16905
#> 
#> [[3]]
#> [1] 11.06892
#> 

# Repeat an unevaluated expression (evaluated with `eval()`)
frepeat(times = 2, expr = quote(rnorm(1)))
#> [[1]]
#> [1] -1.585889
#> 
#> [[2]]
#> [1] -0.3838501
#> 

# Simplify the output to an array
frepeat(times = 3, expr = function() rnorm(1), simplify = TRUE)
#> [1]  0.8648344  0.5612503 -1.4398371

# Monte Carlo simulation: estimate coverage of a 95% CI for sample mean
mc_result <- frepeat(times = 1000, simplify = TRUE, pb = TRUE, ncores = 1,expr = function() {
  sample <- rnorm(30, mean = 0, sd = 1)
  ci <- t.test(sample)$conf.int
  mean(ci[1] <= 0 & 0 <= ci[2])  # check if true mean is inside the interval
})
#>  |                                                  |   0% elapsed=00h 00m 00s, remaining~... |                                                  |   1% elapsed=00h 00m 00s, remaining~00h 00m 00s |=                                                 |   2% elapsed=00h 00m 00s, remaining~00h 00m 00s |=                                                 |   3% elapsed=00h 00m 00s, remaining~00h 00m 00s |==                                                |   4% elapsed=00h 00m 00s, remaining~00h 00m 00s |==                                                |   5% elapsed=00h 00m 00s, remaining~00h 00m 00s |===                                               |   6% elapsed=00h 00m 00s, remaining~00h 00m 00s |===                                               |   7% elapsed=00h 00m 00s, remaining~00h 00m 00s |====                                              |   8% elapsed=00h 00m 00s, remaining~00h 00m 00s |====                                              |   9% elapsed=00h 00m 00s, remaining~00h 00m 00s |=====                                             |  10% elapsed=00h 00m 00s, remaining~00h 00m 00s |=====                                             |  11% elapsed=00h 00m 00s, remaining~00h 00m 00s |======                                            |  12% elapsed=00h 00m 00s, remaining~00h 00m 00s |======                                            |  13% elapsed=00h 00m 00s, remaining~00h 00m 00s |=======                                           |  14% elapsed=00h 00m 00s, remaining~00h 00m 00s |=======                                           |  15% elapsed=00h 00m 00s, remaining~00h 00m 00s |========                                          |  16% elapsed=00h 00m 00s, remaining~00h 00m 00s |========                                          |  17% elapsed=00h 00m 00s, remaining~00h 00m 00s |=========                                         |  18% elapsed=00h 00m 00s, remaining~00h 00m 00s |=========                                         |  19% elapsed=00h 00m 00s, remaining~00h 00m 00s |==========                                        |  20% elapsed=00h 00m 00s, remaining~00h 00m 00s |==========                                        |  21% elapsed=00h 00m 00s, remaining~00h 00m 00s |===========                                       |  22% elapsed=00h 00m 00s, remaining~00h 00m 00s |===========                                       |  23% elapsed=00h 00m 00s, remaining~00h 00m 00s |============                                      |  24% elapsed=00h 00m 00s, remaining~00h 00m 00s |============                                      |  25% elapsed=00h 00m 00s, remaining~00h 00m 00s |=============                                     |  26% elapsed=00h 00m 00s, remaining~00h 00m 00s |=============                                     |  27% elapsed=00h 00m 00s, remaining~00h 00m 00s |==============                                    |  28% elapsed=00h 00m 00s, remaining~00h 00m 00s |==============                                    |  29% elapsed=00h 00m 00s, remaining~00h 00m 00s |===============                                   |  30% elapsed=00h 00m 00s, remaining~00h 00m 00s |===============                                   |  31% elapsed=00h 00m 00s, remaining~00h 00m 00s |================                                  |  32% elapsed=00h 00m 00s, remaining~00h 00m 00s |================                                  |  33% elapsed=00h 00m 00s, remaining~00h 00m 00s |=================                                 |  34% elapsed=00h 00m 00s, remaining~00h 00m 00s |=================                                 |  35% elapsed=00h 00m 00s, remaining~00h 00m 00s |==================                                |  36% elapsed=00h 00m 00s, remaining~00h 00m 00s |==================                                |  37% elapsed=00h 00m 00s, remaining~00h 00m 00s |===================                               |  38% elapsed=00h 00m 00s, remaining~00h 00m 00s |===================                               |  39% elapsed=00h 00m 00s, remaining~00h 00m 00s |====================                              |  40% elapsed=00h 00m 00s, remaining~00h 00m 00s |====================                              |  41% elapsed=00h 00m 00s, remaining~00h 00m 00s |=====================                             |  42% elapsed=00h 00m 00s, remaining~00h 00m 00s |=====================                             |  43% elapsed=00h 00m 00s, remaining~00h 00m 00s |======================                            |  44% elapsed=00h 00m 00s, remaining~00h 00m 00s |======================                            |  45% elapsed=00h 00m 00s, remaining~00h 00m 00s |=======================                           |  46% elapsed=00h 00m 00s, remaining~00h 00m 00s |=======================                           |  47% elapsed=00h 00m 00s, remaining~00h 00m 00s |========================                          |  48% elapsed=00h 00m 00s, remaining~00h 00m 00s |========================                          |  49% elapsed=00h 00m 00s, remaining~00h 00m 00s |=========================                         |  50% elapsed=00h 00m 00s, remaining~00h 00m 00s |=========================                         |  51% elapsed=00h 00m 00s, remaining~00h 00m 00s |==========================                        |  52% elapsed=00h 00m 00s, remaining~00h 00m 00s |==========================                        |  53% elapsed=00h 00m 00s, remaining~00h 00m 00s |===========================                       |  54% elapsed=00h 00m 00s, remaining~00h 00m 00s |===========================                       |  55% elapsed=00h 00m 00s, remaining~00h 00m 00s |============================                      |  56% elapsed=00h 00m 00s, remaining~00h 00m 00s |============================                      |  57% elapsed=00h 00m 00s, remaining~00h 00m 00s |============================                      |  58% elapsed=00h 00m 00s, remaining~00h 00m 00s |=============================                     |  59% elapsed=00h 00m 00s, remaining~00h 00m 00s |==============================                    |  60% elapsed=00h 00m 00s, remaining~00h 00m 00s |==============================                    |  61% elapsed=00h 00m 00s, remaining~00h 00m 00s |===============================                   |  62% elapsed=00h 00m 00s, remaining~00h 00m 00s |===============================                   |  63% elapsed=00h 00m 00s, remaining~00h 00m 00s |================================                  |  64% elapsed=00h 00m 00s, remaining~00h 00m 00s |================================                  |  65% elapsed=00h 00m 00s, remaining~00h 00m 00s |=================================                 |  66% elapsed=00h 00m 00s, remaining~00h 00m 00s |=================================                 |  67% elapsed=00h 00m 00s, remaining~00h 00m 00s |==================================                |  68% elapsed=00h 00m 00s, remaining~00h 00m 00s |==================================                |  69% elapsed=00h 00m 00s, remaining~00h 00m 00s |===================================               |  70% elapsed=00h 00m 00s, remaining~00h 00m 00s |===================================               |  71% elapsed=00h 00m 00s, remaining~00h 00m 00s |====================================              |  72% elapsed=00h 00m 00s, remaining~00h 00m 00s |====================================              |  73% elapsed=00h 00m 00s, remaining~00h 00m 00s |=====================================             |  74% elapsed=00h 00m 00s, remaining~00h 00m 00s |=====================================             |  75% elapsed=00h 00m 00s, remaining~00h 00m 00s |======================================            |  76% elapsed=00h 00m 00s, remaining~00h 00m 00s |======================================            |  77% elapsed=00h 00m 00s, remaining~00h 00m 00s |=======================================           |  78% elapsed=00h 00m 00s, remaining~00h 00m 00s |=======================================           |  79% elapsed=00h 00m 00s, remaining~00h 00m 00s |========================================          |  80% elapsed=00h 00m 00s, remaining~00h 00m 00s |========================================          |  81% elapsed=00h 00m 00s, remaining~00h 00m 00s |=========================================         |  82% elapsed=00h 00m 00s, remaining~00h 00m 00s |=========================================         |  83% elapsed=00h 00m 00s, remaining~00h 00m 00s |==========================================        |  84% elapsed=00h 00m 00s, remaining~00h 00m 00s |==========================================        |  85% elapsed=00h 00m 00s, remaining~00h 00m 00s |===========================================       |  86% elapsed=00h 00m 00s, remaining~00h 00m 00s |===========================================       |  87% elapsed=00h 00m 00s, remaining~00h 00m 00s |============================================      |  88% elapsed=00h 00m 00s, remaining~00h 00m 00s |============================================      |  89% elapsed=00h 00m 00s, remaining~00h 00m 00s |=============================================     |  90% elapsed=00h 00m 00s, remaining~00h 00m 00s |=============================================     |  91% elapsed=00h 00m 00s, remaining~00h 00m 00s |==============================================    |  92% elapsed=00h 00m 00s, remaining~00h 00m 00s |==============================================    |  93% elapsed=00h 00m 00s, remaining~00h 00m 00s |===============================================   |  94% elapsed=00h 00m 00s, remaining~00h 00m 00s |===============================================   |  95% elapsed=00h 00m 00s, remaining~00h 00m 00s |================================================  |  96% elapsed=00h 00m 00s, remaining~00h 00m 00s |================================================  |  97% elapsed=00h 00m 00s, remaining~00h 00m 00s |================================================= |  98% elapsed=00h 00m 00s, remaining~00h 00m 00s |================================================= |  99% elapsed=00h 00m 00s, remaining~00h 00m 00s |==================================================| 100% elapsed=00h 00m 00s, remaining~00h 00m 00s
mean(mc_result)  # estimated coverage
#> [1] 0.946